Added bUnit test project
Added a non async method to fix the displayed number of applications to employer dashboard. Removed unnecessary packages. Added JobifyTestContext so the tests could work on a database. This might need another database that is a copy of the original one or use InMemory. Needs DbContext fix so multiple actions can be used on the database. Added EmployerDashboardTest to test that component.